wouldyouci_back/requirements.txt
amqp==2.6.0
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.3
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.0.6
django-appconf==1.0.4
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-cors-headers==3.2.1
django-elasticsearch-dsl==7.1.1
django-extensions==2.2.9
django-haystack==2.8.1
django-imagekit==4.0.2
django-redis==4.12.1
django-timezone-field==4.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-yasg==1.17.1
elasticsearch==7.7.1
elasticsearch-dsl==7.2.0
haversine==2.2.0
idna==2.9
importlib-metadata==1.6.0
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
joblib==0.15.1
kombu==4.6.9
MarkupSafe==1.1.1
mysqlclient==1.4.6
numpy==1.17.5
packaging==20.3
pandas==1.0.3
pep8==1.7.1
pilkit==2.0
Pillow==7.1.2
PyJWT==1.7.1
PyMySQL==0.9.3
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
redis==3.5.3
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scikit-learn==0.21.3
scikit-surprise==1.1.0
scipy==1.4.1
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
surprise==0.1
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
zipp==3.1.0
